added back indexing, saving like 2x memory by saving y values as i8

+++ learn_bot/learn_bot/libs/hdf5_to_pd.py
@@ -42,7 +42,7 @@ def load_hdf5_to_pd(hdf5_path: Path, selector_df: Optional[pd.DataFrame] = None,
     result_df = pd.concat(partial_dfs, axis=1)
     for col in result_df.columns:
         if str(result_df[col].dtype) == 'bool':
-            result_df[col] = result_df[col].astype(int)
+            result_df[col] = result_df[col].astype('i8')
     return result_df
